Eliminate eta-expansion in hsubst.
The eta-expanion was done by 'ne2nf', and enforced in normal forms
('Nf') by the base type restriction on the neutral ('Ne') argument to
the neutral into normal embedding constructor ('ne').

These changes break the meta theory proofs of course: we necessarily
lose the proofs of decidability of beta-eta equiv, since that was done
by normalization to beta-eta canonical forms, and now we now longer
enforce any kind of eta canonicality.

However, the proofs that the normalizer correctly normalizes,
i.e. that terms are beta-eta equivalent to their normalizations (the
'completeness' theorem in proofs.agda), should be adaptable to this
new definition of normalization.

XXX: would be good to add an example of two beta-eta equal terms with
eta-different normalizations, to convince that I've really removed
the eta-expansion (may not be obvious without studying the code).